feat(DrawingCanvas): add the ability to resize the chessboard when resizing the window
added a method `.calculateSquareSize()` for calculating the size of a square of a chessboard, in order to ensure the correct display of the chessboard regardless of the size of the window.
width and height fields removed as unnecessary.
added `.resize()` method which is responsible for changing the size of the chessboard.
the `.drawPieces()` method now changes the size of the piece depending on the size of the chessboard.
main: added processing of the window resizing event.
main: a `.resize()` function has been added that is responsible for resizing everything in the window.
main: added implementation of a mechanism where when the window size is reduced to less than 512px vertically and horizontally, it does not change size.