feat(ClickHandler): add the ability to cancel focus when clicking on it again
Changed the type for vector indexes from `int` to `size_t`.
Check for repeated clicks on the selected shape has been added to the `.checkClickLeftButtonFromBoard()` method.
From the arguments of the `.analyzeMouseInputFromBoard()` and `.analyzeMouseInput()` methods, those responsible for transmitting click coordinates have been removed; instead, the `event` argument is responsible for transmitting coordinates.