feat(ClassicChessDrawer): add notation to chessboard
Added `FileConverter` File, which is responsible for converting index to char notation and vice versa.
`ClassicChessDrawer`: added the ability to display chess notation on the board taking into account the color of the player’s pieces.
`KeyboardHandler`: added the ability to toggle the visibility of notation when pressing the "C" button.